Add in encoding support in strings.
Now String.v() returns a character array of length (as specified in
the constructor) or NoneObject if the object could not be read.
str(String) returns a string with encoding errors replaced by '?'
and unicode(String) returns a unicode string with encoding errors
replaced by '\ufffd'.

_UNICODE_STRING.v() returns a unicode string or NoneObject if the buffer
length is invalid.  str(_UNICODE_STRING) and unicode(_UNICODE_STRING)
behave like str(String) and unicode(String) except that if there is a
problem with the buffer length or dereferencing the buffer a warning
will be issued.  If the warning can safely be ignored use
str(_UNICODE_STRING or '') or unicode(_UNICODE_STRING or '').

git-svn-id: https://volatility.googlecode.com/svn/trunk@1563 8d5d6628-2090-11de-9909-f37ff7dbbc12