Fix name resolution for types with a CU offset.
When DIEs have a cu_offset, resolving the name via DW_TAG_sibling
has to add the cu_offset to the value of the DW_TAG_sibling to
find the correct element.

Had an instance where the value in a DIE with offset, had a value
that was also valid without adding the required cu_offset,
causing the structure to be incorrectly named and since by chance
the name of the structure was 'long int', it caused a crash as
'long int' is a native type.

ie. following happened, the atomic_t typedef got named
'long int' :

  "long int": [4, {
   "counter": [0, ["int"]]
   }],

Causing a dt display to fail with and obscure error like:

 qemu.dmp 22:50:54> dt target="timespec",offset=0x80925B3C
 -----------------> dt(target="timespec",offset=0x80925B3C)
 [timespec timespec] @ 0x80925b3c
 TypeError: metaclass conflict: the metaclass of a derived class must be a (non-strict) subclass of the metaclasses of all its bases

After this patch, the typedef gets resolved to the correct name:

  "atomic_t": [4, {
   "counter": [0, ["int"]]
   }],