First draft of how multiple ingest queries can be supported. RFC.
For inference-making collectors, it's often useful to be able to make
several ingestion queries instead of just one. Similarly, in those
cases it's important to have all available results in at least the
final call to the inference-making collector.

This is the first attempt at implementing this - I think this is a good
idea, but the code needs to be cleaned up a bit.

Comments welcome.

BUG=
R=scudette@gmail.com

Review URL: https://codereview.appspot.com/164290043