Initial refactor to separate EFILTER from Rekall code.
- Engines remain in the Rekall codebase for the moment. This is mostly
  to make the CL a managable size, and to make it easier to review
  changes to the engines.

- All tests pass locally using the new EFILTER stack.

- Validator no longer checks types. This has always been broken and
  needs a proper implementation based on the new EFILTER types.

- Canonical parser has been renamed to "Dotty" to disambiguate it from
  the name of the project itself.

Next CL will move the engines out of Rekall and into the EFILTER
project, relying on the type interfaces for integration.

Please note that the type system is not meant to be exhaustive at
this point, and is mostly a proof of concept (although it is currently
supporting the object matcher as it inspects entities).

BUG=
R=scudette@gmail.com

Review URL: https://codereview.appspot.com/229510043