Changed logging across the board. Logging now happens via session.logging. A private logging instance is created for every session.
Renderers now have a method called Log that is called every time a
message is to be logged. This is to enable GRR and other projects to
easily trap Rekall logging on a per-session basis.

Added a HoardingLogHandler that stores messages until it's given a
Renderer.

BUG=
R=scudette@gmail.com

Review URL: https://codereview.appspot.com/245810043