Removed Versioneer.
- Versioneer leads to problems because it uses git tags which are hard to manage
  (git pull does not pull them etc). It also requires the source to live in a
  git repo and breaks hard if it is not.

Now there is a special tool version.py which updates static version files before
release. A unique version is calculated from git if it is available and adds the
difference between the last proper release and commit number - but if it is not,
there is a sensible fall back.

R=adam.sindelar@gmail.com

Review URL: https://codereview.appspot.com/295390043 .