fixes wrong indexing with ordinal and takes duplicate functions into account (#413)
The indexing for functions without names has been done with the wrong index: export_directory.Base + i. This is the resulting ordinal number, but not the one used for indexing function pointers. In this case, we just use the already gathered "func" variable.
Furthermore, there are cases where a DLL exports a function twice (e.g. observed for shell32.dll!dllcanunloadnow which is imported by explorer.exe on windows 7). As the ordinal number is used to check for already processed functions, the "functions without names" loop overwrites the entry and the name gets an ordinal number. Using "seen_functions" instead of "seen_ordinals" fixes this:

Results in
0xfff0a600 0x7fefd626810 shell32.dll!dllcanunloadnow
instead of
0xfff0a600 0x7fefd626810 shell32.dll!259